--- domain/user.py
@@ -1,8 +1,0 @@
-class User:
-    def __init__(self, user_id, user_pw, user_email, user_nickname, user_profileimg):
-        self.user_id = user_id
-        self.user_pw = user_pw
-        self.user_email = user_email
-        self.user_nickname = user_nickname
-        self.user_profileimg = user_profileimg
-
